+++ stdnum/eu/nace/v2_1/__init__.py
@@ -1,0 +1,10 @@
+from .nace import *
+
+
+__all__ = (list(getattr(nace, "__all__", []))or [n for n in dir(nace) if not n.startswith("_")])
+
+# remove nace as submodule reference
+try:
+    del nace
+except NameError:
+    pass
